+++ index.js
@@ -1,4 +1,4 @@
-const { truePrice } = require('./truePrice') 
+const {truePrice} = require('./truePrice');
 
 // const filters = { showOnly: { completed: true, sold: true, dealsAndSavings: true }}
 // const options = {
@@ -6,9 +6,9 @@ const { truePrice } = require('./truePrice')
 //     verbose: true,
 //     filters: filters,
 // }
-// 
+//
 // truePrice.fetchAveragePrice(options).then((res) => console.log(res));
 
 module.exports = {
-    truePrice: truePrice
-}
+  truePrice: truePrice,
+};